lv1 lv2 나누기하면 결과값 실수로 받는 기능 추가 lv3 패키지 및 class 추가 class 내용은 lv2 내용과 동일

+++ src/lv1/Main.java
@@ -40,7 +40,7 @@ public class Main {
                             break;
                         case "/" :
                             if(num2 != 0){
-                                System.out.println("결과 : " + (num1 / num2));
+                                System.out.println("결과 : " + (double)(num1 / num2));
                             }
                             else {
                                 System.out.println("0으로 나눌 수 없습니다.");
